use function to transform coordinates to new layer
