+++ src/interface/src/app/map/map.helper.ts
@@ -1,4 +1,10 @@
-import { Feature, FeatureCollection, MultiPolygon, Polygon } from 'geojson';
+import {
+  Feature,
+  FeatureCollection,
+  MultiPolygon,
+  Polygon,
+  Position,
+} from 'geojson';
 import * as L from 'leaflet';
 import booleanWithin from '@turf/boolean-within';
 import booleanIntersects from '@turf/boolean-intersects';
@@ -172,4 +178,15 @@ export function defaultMapConfigsDictionary(): Record<Region, MapConfig[]> {
   };
 }
 
-export function transformMultiPolygonToPolygon() {}
+export function transformCoordToLayer(coord: Position[][]) {
+  const newFeature: GeoJSON.Feature = {
+    type: 'Feature',
+    geometry: {
+      type: 'Polygon',
+      coordinates: coord,
+    },
+    properties: {},
+  };
+
+  return new L.GeoJSON(newFeature);
+}
